rag_system/backend/requirements.txt
# FastAPI Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Google Cloud
google-cloud-aiplatform==1.71.1
google-cloud-discoveryengine==0.13.10
google-cloud-firestore==2.19.0
firebase-admin==6.6.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
gspread==6.1.4
oauth2client==4.1.3

# Vector & Search
numpy==2.1.3
scikit-learn==1.5.2
rank-bm25==0.2.2

# Async & HTTP
httpx==0.27.2
aiohttp==3.11.2
sse-starlette==2.1.3

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
orjson==3.10.10

# Monitoring & Logging
structlog==24.4.0
langsmith>=0.1.0
langchain-core>=0.1.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
ruff==0.7.2
mypy==1.13.0

# Types
types-requests==2.32.0.20241016
